% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStrStage.R
\name{calcStrStage}
\alias{calcStrStage}
\title{calcStrStage: calculate structural stage for each plot}
\usage{
calcStrStage(
  park = "all",
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  from = 2006,
  to = 2019,
  output,
  speciesType = c("all", "native", "exotic"),
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one park at a time. Acceptable options are:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots, such as deer exclosures
\describe{
\item{"VS"}{Default. Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as deer exclosures and bonus plots}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1,3), for example.}

\item{from}{Year to start analysis, ranging from 2006-2019}

\item{to}{Year to stop analysis, ranging from 2006-2019}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}
}
\value{
returns a dataframe with structural stage and metrics used to assign stages to plots.
}
\description{
This function calculates structural stage metric from Ecological Integrity Scorecard,
which assigns Pole, Mature, Late Successional, or Mosaic (i.e., none of the above) to each plot based
on the percent of live basal area of canopy trees in pole, mature and large size classes. Plots must be
closed-canopy forest to be classified for this metric. Therefore plots classified as Woodlands (ACAD only)
or Early successional (SARA only) in the field are automartically assigned those classes in the calculation.
Must run importData first.
}
\examples{
importData() #imports using default odbc
stage_df <- calcStrStage(park = 'MABI', from = 2016, to = 2019)


}
