% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCSV.R
\name{importCSV}
\alias{importCSV}
\title{importCSV: Import NETN forest data that are formated as .csv files.}
\usage{
importCSV(path = NA)
}
\arguments{
\item{path}{Quoted path of folder containing tables.}
}
\value{
Assigns database tables to global environment
}
\description{
This function imports exported CSV tables from the NETN Forest Database, and
assigns the tables to the global environment with names that functions in this package
depend on.
}
\examples{
importCSV("C:/Data/forest_csvs")

}
