% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLocEvent.R
\name{joinLocEvent}
\alias{joinLocEvent}
\title{joinLocEvent: merges Location and Event level data with options for filtering.}
\usage{
joinLocEvent(
  park = "all",
  from = 2006,
  to = 2019,
  QAQC = FALSE,
  rejected = FALSE,
  panels = 1:4,
  locType = "VS",
  eventType = c("complete", "all"),
  output = "short",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one park at a time. Acceptable options are:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006-2019}

\item{to}{Year to stop analysis, ranging from 2006-2019}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{rejected}{Allows you to remove (FALSE) or include (TRUE) rejected plots.
\describe{
\item{FALSE}{Default. Only returns plots that were not rejected.}
\item{TRUE}{returns all records}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1,3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots, such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as deer exclosures and bonus plots}}}

\item{eventType}{Allows you to only include complete sampling events, or to include all sampling events
\describe{
\item{"complete"}{Only include sampling events for a plot that are complete.}
\item{"all}{Include all plots with an Event_ID, including plots that are missing all data associated with that event (eg ACAD-029.2010).}
}}
}
\value{
returns a dataframe with location and visit events
}
\description{
This function combines location and event data. Must run importData first.
}
\examples{
importCSV('./forest_csvs')
# Select most recent survey of data from WEFA
WEFA_data <- joinLocEvent(park = 'WEFA', panels = c(2,4), from = 2015, to = 2018)

# Select data from cycle 3
cycle3 <- joinLocEvent(from = 2014, to = 2017) # all parks is default

# Select data from plots that had a QA/QC event in ACAD in 2018
ACAD_data<-joinLocEvent(park = 'ACAD', QAQC = T, from = 2018)
QAQC_plots<-ACAD_data$Plot_Name[which(ACAD_data$Event_QAQC==TRUE)]
ACAD_QAQC<-ACAD_data \%>\% filter(Plot_Name \%in\% QAQC_plots) \%>\% droplevels()

}
