% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinMicroShrubData.R
\name{joinMicroShrubData}
\alias{joinMicroShrubData}
\title{joinMicroShrubData: compiles shrub data collected in microplots}
\usage{
joinMicroShrubData(
  speciesType = c("all", "native", "exotic"),
  numMicros = 3,
  park = "all",
  from = 2007,
  to = 2018,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{numMicros}{Allows you to select 1, 2, or 3 microplots of data to summarize}
}
\value{
returns a dataframe with shrub data collected in microplots
}
\description{
This function combines shrub stem counts (cycle 1) and percent cover (cycle 2+) data from microplots. Must run importData first.
}
\examples{
importData()
# native shrubs in MORR all years
native_shrubs <- joinMicroShrubData(park ='MORR', speciesType = 'native')

# all parks with exotic shrubs in most recent survey
exotic_shrubs <- joinMicroShrubData(from = 2015, to = 2018, speciesType = 'exotic')

}
