% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinQuadData.R
\name{joinQuadData}
\alias{joinQuadData}
\title{joinQuadData: compiles quadrat species data}
\usage{
joinQuadData(
  speciesType = c("all", "native", "exotic", "invasive"),
  park = "all",
  from = 2006,
  to = 2018,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}
}
\value{
Returns a dataframe with cover class midpoints for each quadrat and includes guild for each species.
}
\description{
This function combines quadrat species data with species names and allows you to filter on species types, park, years, and visit type. Note that the Shrub guild also includes woody vine species.
}
\examples{
importData()
# compile quadrat data for invasive species in SARA for all years
SARA_quads <- joinQuadData(park = 'SARA', speciesType = 'invasive')

# compile native species only for all parks in most recent survey
native_quads <- joinQuadData(speciesType = 'native', from = 2015, to = 2018)

}
