% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinRegenData.R
\name{joinRegenData}
\alias{joinRegenData}
\title{joinRegenData: compiles seedling and sapling data}
\usage{
joinRegenData(
  speciesType = c("all", "native", "exotic"),
  canopyForm = c("canopy", "all"),
  numMicros = 3,
  units = c("micro", "ha", "acres"),
  park = "all",
  from = 2006,
  to = 2018,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{canopyForm}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Returns all species, including low canopy species.}
\item{"canopy"}{Default. Returns canopy-forming species only}
}}

\item{numMicros}{Allows you to select 1, 2, or 3 microplots of data to summarize}

\item{units}{Calculates seedling and sapling densities based on different units.
\describe{
\item{"micro"}{Default. Returns seedling and sapling densities per microplot.}
\item{"ha"}{Returns seedling and sapling densities per hectare}
\item{"acres"}{Returns densities per acre}
}}
}
\value{
returns a dataframe with seedling and sapling densities, and stocking index
}
\description{
This function combines seedling and sapling data, and calculates stocking index. Must run importData first.
}
\examples{
importCSV('./forest_csvs/')
# compile seedling and sapling data for all parks and all species in most recent cycle
regen_data <- joinRegenData(canopyForm = 'all', from = 2015, to = 2018)

# compile regen data for only canopy-forming (default) and native species in SAGA for all years
SAGA_regen <- joinRegenData(park = 'SAGA', speciesType = 'native')

# compile only 1 microplot of data for ACAD native canopy-forming species for all but first year
ACAD_regen_m1 <- joinRegenData(park = 'ACAD', speciesType = 'native', numMicros = 1, from = 2007)

}
