% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSoilLabData.R
\name{joinSoilLabData}
\alias{joinSoilLabData}
\title{joinSoilLabData: compile soil chemistry data by horizon.}
\usage{
joinSoilLabData(
  park = "all",
  from = 2007,
  to = 2018,
  layers = c("all", "O", "A"),
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  ...
)
}
\value{
returns a dataframe containing each plot and visit with soil chemistry data for each horizon on a plot
Plots that weren't sampled during a given cycle are not returned. Horizon depths are averaged across samples.
}
\description{
This function verifies whether O and A horizons were named corrected based on % Total Carbon.
For duplicate horizons on a plot, chemistry variables are corrected using weighted averages, with sample depth
as the weight. Must run importData first. Note that Earthworms are summarized in joinStandData().
Note that data starts at 2007 because 2006 methods were pretty different.
}
\examples{
importData() #imports using default odbc
# join only O horizon data for most recent cycle in ACAD. Note soil chemistry data are typically a year behind plot data.
soil_ACAD_O <- joinSoilLabData(park = 'ACAD', from = 2015, to = 2018, layers = 'O')

# join all park data from all layers and all years
soil_df_all <- joinSoilLabData(from = 2007, to = 2018, layers = 'all')

}
