% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSoilSampleData.R
\name{joinSoilSampleData}
\alias{joinSoilSampleData}
\title{joinSoilSampleData: compile corrected soil sample data.}
\usage{
joinSoilSampleData(
  park = "all",
  from = 2007,
  to = 2018,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  layers = "all",
  ...
)
}
\value{
returns a dataframe containing each plot and visit with soil sample data.Tot_Samp_cm is the total depth of O and A sampled.
Litter is not included in total depth calculation. Note soil chemistry data are typically a year behind plot data,
so correcteed soil horizon depths will also be a year behind. Plots that weren't sampled during a given cycle are
not returned. Horizon depths are averaged across samples of the same horizon type.
}
\description{
This function verifies whether O and A horizons were named corrected based on % Total Carbon using
the joinSoilLabData() function, then compiles sample depth by horizon. Must run importData first.
Note that Earthworms are summarized in joinStandData(). Note that data starts at 2007 because 2006 methods were pretty different.
}
\examples{
importData() #imports using default odbc
# join horizon dpeth data for most recent cycle in ACAD.
soil_ACAD_O <- joinSoilSampleData(park = 'ACAD', from = 2015, to = 2018)

}
