% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumQuadGuilds.R
\name{sumQuadGuilds}
\alias{sumQuadGuilds}
\title{sumQuadGuilds: summarizes quadrat species data by guilds}
\usage{
sumQuadGuilds(
  speciesType = c("native", "exotic", "invasive", "all"),
  park = "all",
  from = 2006,
  to = 2019,
  splitHerb = TRUE,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  output,
  ...
)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Returns all species.}
\item{"native"}{Default. Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the NETN Indicator Invasive List}
}}

\item{splitHerb}{TRUE/FALSE. If TRUE (default), allows you to split the herbaceous group into forb and fern. If FALSE,
then resulting data frame will be summarised for tree, shrub, herbaceous, and graminoid guilds.}
}
\value{
Returns a dataframe with average quadrat cover, percent quadrat frequency and quadrat frequency count for tree,shrub/vine,herbaceous,and graminoid. Data are sither summarized for all species, native only, exotic only, or invasive only.
}
\description{
This function summarizes output from joinQuadData and calculates average cover and quadrat frequency for each guild.
Average cover is corrected for number of quadrats sampled. Guilds are tree, shrub, forb, fern, and graminoid. If herbaceous guild
is split, then cover of ferns does not overlap with cover of herbaceous. If herbaceous guild is not split, then cover of herbaceous
guild includes fern and other herbaceous (but not graminoid) species cover.
}
\examples{
importData()

# compile invasive quad data for all parks and most recent survey. Keep ferns in with herbs
inv_guilds <- sumQuadGuilds(speciesType = 'invasive', from = 2015, to = 2018, splitHerb = FALSE)

# compile native quad data for more recent survey in ACAD, with ferns and forbs split in separate guilds
ACAD_guilds <- sumQuadGuilds(speciesType = 'native', from = 2015, to = 2018, splitHerb = TRUE)

}
