% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTreeDBHDist.R
\name{sumTreeDBHDist}
\alias{sumTreeDBHDist}
\title{sumTreeDBHDist: calculates DBH distribution of trees}
\usage{
sumTreeDBHDist(
  status = c("all", "live", "dead"),
  speciesType = c("all", "native", "exotic"),
  units = c("density", "ba"),
  park = "all",
  from = 2006,
  to = 2019,
  QAQC = FALSE,
  locType = "VS",
  panels = 1:4,
  dist_m = NA,
  output,
  ...
)
}
\arguments{
\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"all"}{Includes all standing trees}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{units}{Allows you to choose which metric to calculate: basal area or stem density
\describe{
\item{"density"}{Default. Returns stems/ha}
\item{"ba"}{Returns basal area in sq.m/ha}
}}

\item{dist_m}{Filter trees by a distance that is less than or equal to the specified distance in meters
of the tree to the center of the plot. If no distance is specified, then all trees will be selected. For
example, to select an area of trees that is 100 square meters in area, use a distance of 5.64m.}
}
\value{
returns a dataframe with one row for each plot and either density or BA
}
\description{
This function calculates DBH distribution by 10cm size classes. Must run importData first.
}
\examples{
importData()
tree_diam_dist <-sumTreeDBHDist(park = 'MORR', speciesType = 'native', from = 2016, to = 2019, units = 'ba')
head(tree_diam_dist)

}
