% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_FHM.R
\name{theme_FHM}
\alias{theme_FHM}
\title{theme_FHM: custom ggplot2 theme for forestNETN}
\usage{
theme_FHM()
}
\value{
This function must be used in conjunction with a ggplot object
}
\description{
This function customizes a theme for plotting NETN forest data, including removing the
default panel grids from ggplot2 figures.
}
\examples{
importData()
 size_dist <- ggplot(mean_dens, aes(x = size_class, y = stems_ha_mean))+
     geom_bar(stat = 'identity', fill = 'CadetBlue', position = position_dodge())+
     geom_errorbar(aes(ymin=stems_ha_mean-stems_ha_se, ymax=stems_ha_mean+stems_ha_se, x=size_class),
                   color='DimGrey', width=0.2,position=position_dodge(0.9))+
     facet_wrap(vars(cycle_fac), ncol=3, labeller=as_labeller(cycle_names))+
     labs(x='Regeneration Size Distribution', y='stems/ha')+
     scale_x_discrete(labels=c('15-30cm', '30-100cm', '100-150cm', '>150cm','1-10cm DBH'))+
     theme_FVM()

}
