% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMapPrep.R
\name{treeMapPrep}
\alias{treeMapPrep}
\title{treeMapPrep: prepares tree data for treeMap (internal function)}
\usage{
treeMapPrep(df)
}
\value{
Returns a map of trees on a given plot
}
\description{
This function converts tree distance and azimuth values to coordinates and plots the coordinates of live or
dead trees, then simplifies status codes for the map. Works best with data.frame derived
from joinLocEvent and joinTreeData. Requires a data.frame with the following fields:
 Plot_Name, Tree_Number_NETN, Distance, Azimuth, Orientation, Status_ID, and DBH. This function is used in treeMap().
}
